Move the interfaces to the consumers
