+++ Product Observer/messageBroker/messageBroker_interface.go
@@ -1,8 +1,8 @@
 package messageBroker
 
 // MessageBroker interface
-type MessageBroker interface {
-	PublishMessages(exchange string, queueName string, messages ...string) error
-	ConsumeMessages(queueName string, handler func(message string)) error
-	Close() error
-}
+// type MessageBroker interface {
+// 	PublishMessages(exchange string, queueName string, messages ...string) error
+// 	ConsumeMessages(queueName string, handler func(message string)) error
+// 	Close() error
+// }
