+++ publisher/database/database_Interface.go
@@ -1,13 +1,12 @@
 package database
 
 //TODO: Remove this Package Make it puplic
-import "database/sql"
 
-type DbConnector interface {
-	// start() error
-	Insert(tableName string, argsKeys []string, argsVals []string) error
-	CloseDB() error
+// type DbConnector interface {
+// 	// start() error
+// 	Insert(tableName string, argsKeys []string, argsVals []string) error
+// 	CloseDB() error
 
-	//TODO: Remove SQL Package Make it puplic
-	SelectById(tableName string, id int) (*sql.Rows, error)
-}
+// 	//TODO: Remove SQL Package Make it puplic
+// 	SelectById(tableName string, id int) (*sql.Rows, error)
+// }
