+++ publisher/product/product_Interface.go
@@ -1,6 +1,6 @@
 package product
 
-type ProductManager interface {
-	Add(product *Product, queueName string) error
-	GetByID(id int) (*Product, error)
-}
+// type ProductManager interface {
+// 	Add(product *Product, queueName string) error
+// 	GetByID(id int) (*Product, error)
+// }
